refactor(kernel): move the read-only sqlite ledger reconciliation out of the store module
reconcileSqliteEvents opens the database read-only and never writes; keeping it
in sqlite-event-store.ts made the bypass-write gate treat its two SELECTs as
governed writes. It now lives in sqlite-ledger-reconcile.ts, a pure move.
