+++ packages/kernel/src/composition/index.ts
@@ -13,7 +13,8 @@ export {
 export type { WalRecoveryProgress } from "../store/wal-shadow-event-store.ts";
 export { ledgerGitPath, resolveLedgerGitLayout } from "../store/ledger-git-layout.ts";
 export { eventShapeMigrations, runEventShapeMigration } from "../store/event-shape-migration.ts";
-export { migrateEventsToSqlite, openSqliteEventStore, reconcileSqliteEvents } from "../store/sqlite-event-store.ts";
+export { migrateEventsToSqlite, openSqliteEventStore } from "../store/sqlite-event-store.ts";
+export { reconcileSqliteEvents } from "../store/sqlite-ledger-reconcile.ts";
 export { runDispatchRecordMigration } from "../store/dispatch-record-migration.ts";
 export type { DispatchRecordLeaseSettlement } from "../store/dispatch-record-migration.ts";
 export { resolveRetirableDocument } from "../store/ledger-document.ts";
