+++ packages/kernel/src/store/index.ts
@@ -1,5 +1,6 @@
 export * from "../integrity/stable-hash.ts";
 export * from "./entity-store.ts";
 export * from "./local-version-control-system.ts";
-export { migrateEventsToSqlite, openSqliteEventStore, reconcileSqliteEvents } from "./sqlite-event-store.ts";
+export { migrateEventsToSqlite, openSqliteEventStore } from "./sqlite-event-store.ts";
+export { reconcileSqliteEvents } from "./sqlite-ledger-reconcile.ts";
 export * from "./task-event-store.ts";
